implement react redux to weather
