+++ Challenge/index.android.js
@@ -1,8 +1,7 @@
 
 import React, { Component } from 'react'
 import {
-  AppRegistry,
-  View
+  AppRegistry
 } from 'react-native'
 import {
   StackNavigator
@@ -10,16 +9,6 @@ import {
 
 import Landing from './src/screens/Landing'
 import Weather from './src/screens/Weather'
-
-class Challenge extends Component {
-  render() {
-    return (
-      <View>
-        <Weather />
-      </View>
-    );
-  }
-}
 
 const App = StackNavigator({
   Landing: { screen: Landing },
